+++ Assets/OxGKit/InfiniteScrollView/Scripts/Runtime/Core/CellData/InfiniteCellData.cs
@@ -6,13 +6,18 @@ namespace OxGKit.InfiniteScrollView
     public class InfiniteCellData : IDisposable
     {
         public int index { get; internal set; }
+
+        /// <summary>
+        /// Cell width and height (x, y)
+        /// </summary>
         public Vector2 cellSize;
+
+        /// <summary>
+        /// Data required to render this cell
+        /// </summary>
         public object data;
 
-        public InfiniteCellData()
-        {
-
-        }
+        public InfiniteCellData() { }
 
         public InfiniteCellData(Vector2 cellSize)
         {