+++ Assets/OxGKit/InfiniteScrollView/Scripts/Runtime/Core/ScrollView/HorizontalGridInfiniteScrollView.cs
@@ -43,7 +43,8 @@ namespace OxGKit.InfiniteScrollView
                             if (index >= this._dataList.Count)
                                 break;
                             var visibleRange = new Vector2(contentWidth, contentWidth + this._dataList[index].cellSize.x);
-                            if (visibleRange.y < viewportRange.x || visibleRange.x > viewportRange.y)
+                            if (visibleRange.y < viewportRange.x ||
+                                visibleRange.x > viewportRange.y)
                             {
                                 this.RecycleCell(index);
                             }
@@ -61,7 +62,8 @@ namespace OxGKit.InfiniteScrollView
                                 index >= this._dataList.Count)
                                 break;
                             var visibleRange = new Vector2(contentWidth, contentWidth + this._dataList[index].cellSize.x);
-                            if (visibleRange.y < viewportRange.x || visibleRange.x > viewportRange.y)
+                            if (visibleRange.y < viewportRange.x ||
+                                visibleRange.x > viewportRange.y)
                             {
                                 this.RecycleCell(index);
                             }
@@ -85,22 +87,32 @@ namespace OxGKit.InfiniteScrollView
                             if (index >= this._dataList.Count)
                                 break;
                             var visibleRange = new Vector2(contentWidth, contentWidth + this._dataList[index].cellSize.x);
-                            if (visibleRange.y >= viewportRange.x && visibleRange.x <= viewportRange.y)
+                            if (visibleRange.y >= viewportRange.x &&
+                                visibleRange.x <= viewportRange.y)
                             {
                                 // Calculate visible count
                                 this.visibleCount++;
-                                if (this.visibleCount % this.rowCount == 0) lastVisibleWidth = visibleRange.y;
+                                if (this.visibleCount % this.rowCount == 0)
+                                    lastVisibleWidth = visibleRange.y;
 
                                 InfiniteCell cell = null;
                                 if (this._cellList[index] == null)
                                 {
-                                    if (this._cellPool.Count > 0) cell = this._cellPool.Dequeue();
-                                    else Logging.PrintError<Logger>("The cell display error occurred, not enough cells in the cell pool!!!");
+                                    // Get cell from pool
+                                    if (this._cellPool.Count > 0)
+                                        cell = this._cellPool.Dequeue();
+                                    else
+                                        Logging.PrintError<Logger>("The cell display error occurred, not enough cells in the cell pool!!!");
                                 }
+
                                 // Check cell direction pivot
                                 float dirCoeff = 1f;
-                                if (cell != null) dirCoeff = cell.rectTransform.pivot.x > 0 ? -1f : 1f;
+                                if (cell != null)
+                                    dirCoeff = cell.rectTransform.pivot.x > 0 ? -1f : 1f;
+
+                                // Begin setup cell to display
                                 this.SetupCell(cell, index, new Vector2(contentWidth * dirCoeff, (this._dataList[index].cellSize.y + this.spacing.y) * -j + -(this.padding.top - this.padding.bottom)));
+
                                 if (visibleRange.y >= viewportRange.x)
                                     this._cellList[index]?.transform.SetAsLastSibling();
                                 else
@@ -117,25 +129,35 @@ namespace OxGKit.InfiniteScrollView
                         {
                             int index = i - j;
                             if (index < 0 ||
-                            index >= this._dataList.Count)
+                                index >= this._dataList.Count)
                                 break;
                             var visibleRange = new Vector2(contentWidth, contentWidth + this._dataList[index].cellSize.x);
-                            if (visibleRange.y >= viewportRange.x && visibleRange.x <= viewportRange.y)
+                            if (visibleRange.y >= viewportRange.x &&
+                                visibleRange.x <= viewportRange.y)
                             {
                                 // Calculate visible count
                                 this.visibleCount++;
-                                if (this.visibleCount % this.rowCount == 0) lastVisibleWidth = visibleRange.y;
+                                if (this.visibleCount % this.rowCount == 0)
+                                    lastVisibleWidth = visibleRange.y;
 
                                 InfiniteCell cell = null;
                                 if (this._cellList[index] == null)
                                 {
-                                    if (this._cellPool.Count > 0) cell = this._cellPool.Dequeue();
-                                    else Logging.PrintError<Logger>("The cell display error occurred, not enough cells in the cell pool!!!");
+                                    // Get cell from pool
+                                    if (this._cellPool.Count > 0)
+                                        cell = this._cellPool.Dequeue();
+                                    else Logging.
+                                            PrintError<Logger>("The cell display error occurred, not enough cells in the cell pool!!!");
                                 }
+
                                 // Check cell direction pivot
                                 float dirCoeff = 1f;
-                                if (cell != null) dirCoeff = cell.rectTransform.pivot.x > 0 ? -1f : 1f;
+                                if (cell != null)
+                                    dirCoeff = cell.rectTransform.pivot.x > 0 ? -1f : 1f;
+
+                                // Begin setup cell to display
                                 this.SetupCell(cell, index, new Vector2(contentWidth * dirCoeff, (this._dataList[index].cellSize.y + this.spacing.y) * -j + -(this.padding.top - this.padding.bottom)));
+
                                 if (visibleRange.y >= viewportRange.x)
                                     this._cellList[index]?.transform.SetAsLastSibling();
                                 else
@@ -177,6 +199,7 @@ namespace OxGKit.InfiniteScrollView
         {
             if (!this.IsInitialized())
                 return;
+
             if (index >= this._dataList.Count ||
                 index < 0)
                 return;
@@ -219,25 +242,27 @@ namespace OxGKit.InfiniteScrollView
         #endregion
 
         #region Sealed Override
-        public sealed override void Refresh(bool disabledRefreshCells = false)
+        public sealed override void Refresh(bool refreshOnNextScroll = false, bool recycleActiveCells = false)
         {
-            if (!this.IsInitialized()) return;
+            if (!this.IsInitialized())
+                return;
 
             if (this.scrollRect.viewport.rect.width == 0)
             {
-                this.DoDelayRefresh(disabledRefreshCells).Forget();
+                this.DoDelayRefresh(refreshOnNextScroll, recycleActiveCells).Forget();
             }
             else
             {
-                this.DoRefresh(disabledRefreshCells);
+                this.DoRefresh(refreshOnNextScroll, recycleActiveCells);
             }
         }
 
-        protected sealed override void DoRefresh(bool disabledRefreshCells)
+        protected sealed override void DoRefresh(bool refreshOnNextScroll, bool recycleActiveCells)
         {
-            if (this.scrollRect == null) return;
+            if (this.scrollRect == null)
+                return;
 
-            if (!disabledRefreshCells)
+            if (!refreshOnNextScroll)
             {
                 // Refresh content size
                 float width = this.padding.left;
@@ -248,11 +273,8 @@ namespace OxGKit.InfiniteScrollView
                 width += this.padding.right;
                 this.scrollRect.content.sizeDelta = new Vector2(width, this.scrollRect.content.sizeDelta.y);
 
-                // Recycle all cells first
-                for (int i = 0; i < this._cellList.Count; i++)
-                {
-                    this.RecycleCell(i);
-                }
+                if (recycleActiveCells)
+                    this.RecycleActiveCells();
 
                 // Refresh cells view
                 this.DoRefreshVisibleCells();
@@ -261,13 +283,14 @@ namespace OxGKit.InfiniteScrollView
                 this.onRefreshed?.Invoke();
             }
             // Mark flag for refresh at next scrolling
-            else this._disabledRefreshCells = true;
+            else
+                this._refreshOnNextScroll = true;
         }
 
-        protected sealed override async UniTask DoDelayRefresh(bool disabledRefreshCells)
+        protected sealed override async UniTask DoDelayRefresh(bool refreshOnNextScroll, bool recycleActiveCells)
         {
             await UniTask.Yield(PlayerLoopTiming.LastPostLateUpdate);
-            this.DoRefresh(disabledRefreshCells);
+            this.DoRefresh(refreshOnNextScroll, recycleActiveCells);
         }
         #endregion
     }